feat: refactor Redis connection handling with centralized configuration and validation
